requirements.txt
langchain
openai
ipykernel
python-dotenv
huggingface-hub
streamlit